+++ coupon/src/main/java/com/sparta/coupon/application/service/CouponService.java
@@ -3,7 +3,9 @@ package com.sparta.coupon.application.service;
 
 import static com.sparta.coupon.application.exception.Error.NOT_FOUND_COUPON;
 
-import com.sparta.coupon.application.dto.request.CouponDto;
+import com.sparta.coupon.application.dto.request.CreateCouponRequestDto;
+import com.sparta.coupon.application.dto.response.GetCouponDetailResponseDto;
+import com.sparta.coupon.application.dto.response.GetCouponResponseDto;
 import com.sparta.coupon.application.exception.CouponException;
 import com.sparta.coupon.infrastructure.repository.CouponRepository;
 import com.sparta.coupon.model.core.Coupon;
@@ -21,32 +23,32 @@ public class CouponService {
     private final CouponRepository couponRepository;
 
     @Transactional
-    public CouponDto.GetResponse createCoupon(CouponDto.CreateRequest requestDto) {
+    public GetCouponResponseDto createCoupon(CreateCouponRequestDto requestDto) {
 
         Coupon coupon = requestDto.toEntity();
         couponRepository.save(coupon);
 
-        return CouponDto.GetResponse.from(coupon);
+        return GetCouponResponseDto.from(coupon);
 
     }
 
     @Transactional(readOnly = true)
-    public CouponDto.GetDetailResponse getCoupon(Long id) {
+    public GetCouponDetailResponseDto getCoupon(Long id) {
 
         Coupon coupon = couponRepository.findByIdAndIsDeletedFalse(id)
                 .orElseThrow(() -> new CouponException(NOT_FOUND_COUPON, HttpStatus.NOT_FOUND));
 
-        return CouponDto.GetDetailResponse.from(coupon);
+        return GetCouponDetailResponseDto.from(coupon);
     }
 
     @Transactional(readOnly = true)
-    public List<CouponDto.GetDetailResponse> getAllCoupons() {
+    public List<GetCouponDetailResponseDto> getAllCoupons() {
 
         List<Coupon> coupons =  couponRepository.findByIsDeletedFalse()
                 .orElseThrow(() -> new CouponException(NOT_FOUND_COUPON, HttpStatus.NOT_FOUND));
 
         return  coupons.stream()
-                .map(CouponDto.GetDetailResponse::from)
+                .map(GetCouponDetailResponseDto::from)
                 .collect(Collectors.toList());
     }
